+++ transform/models/trusted_zone/rndts/dd_sortants.sql
@@ -42,7 +42,7 @@ select
     code_dechet,
     dechet_pop,
     denomination_usuelle,
-    quantite,
+    quantite::numeric,
     unite,
     date_expedition::date,
     status
